Update README.md to outline steps for deployment
Readme file was updated to refelct changes to deployment and give
current and future teams a guideline of where to start with coding and
deployment. System Architecture and Testing information will be added
later on as that information becomes available. Deployment steps may
change as we progress to live deployment.