CW-42 Delete previous error handling.
Error handling of database previously would kill the application,
shutting down the server and stop serving the frontend. If this was
continued into deployment, then this would be a major problem with a minor
fix that might not be found. This was found when testing the database
queries as there was no way to retrieve errors without propogating up.
Errors will be handled in a fix to the APIs later down the line.