Refactor database queries and add validation.
Due to the fact that the database queries were becoming one line code as
I refactored, the decision was made to absorb that code into the API
handles. This may be changed back due to abstraction and security but as
of now it is cleaner code.

A validation pacakage was also introduced to reduce the number of lines
used for validating.