Optimse docker workflow for development.
The dockerfiles were optimised for production and not for development.
Hence it was difficult for the team to use the dockerfiles to run as a
development environment. These changes will hopefully eliminate the need
to rebuild an image everytime a change is made. Testing will begin in
the testing branch.