CW-40 Added faq_test.go (#41)
* added faq_test.go to test faq.go via testing package

* Remove boboeee as codeowner.

To streamline PR reviews so that only concerned parties need to give the
review. Members will still be informed of a PR through the slack
integration.

* added faq_test.go to test faq.go via testing package

* CW-40 Remove assertStatus function

Due to the fact that all test are in the same package the function was
causing the compiler to throw a redeclaration error. This needs to be
addressed with a general refactoring of backend code.

Co-authored-by: Paula Tennent <pjtennent@gmail.com>