CW-71 Separate docker file functionality.
To streamline development, Vue serve will be used in development. This
means that deployment dockerfiles and vue configurations will need to be
created at the time of deployment for the purposes fo production.

Leveraging caches, separate dockerfiles were created for the frontend
and backend. So current dockerfiles in dev branch surfice for deployment
but code will need to be changed if doing so.