CW-44 Compose emails and send them using Mailjet (#47)
* Created a handle to deal with enquiries

'Contact Us' and 'Sponsorship Request' forms have the same fields (name, email, message). They will use the same handler but different endpoints to distinguish the emails that these enquiries must be forwarded to.

* Added starting code from Gmail API documentation

The Gmail API documentation found at https://developers.google.com/gmail/api/quickstart/go

* Created an init function to connect with the Gmail API

InitMailingClient() creates a client instance to the Gmail API and keeps it private to the rest of the server. The server can only interact with it by calling another function to send emails.

* Developed a function to send emails from a pre-authorised account

SendEmail() takes in the fields from the enquiry form; it then composes and sends an email to a target address.

* Updated module files to include new dependencies

* Mailing init function exits with a fatal error if it fails

* Refactor handleEnquiry into mailing.go

Refactored to encapsulate all mail related functions together in a separate file.

* Use structs for validation

The purpose of validating thorugh structs is to simplify and unify the verification process.

* Implement the Message struct developed in CW-34

mailing.go is only interfaced with using InitMailingClient() and HandleEnquiry()

* Establish simple SMTP connection to Gmail to send emails

Commented out previous Gmail API approach and modified functions to establish a SMTP connection to Gmail enabling only the sending of emails from an authenticated account (csesoc@csesoc.org.au)

* Extract message formatting to a helper function

composeEmail formats the Message struct and a target email to be of RFC 822-style ready to be sent through SMTP

* Use a library to obtain the password for the SMTP authentication

The password is obtained from stdin (with echo off) upon server initialisation

* Fix error checking in getPassword()

* CW-62 Unit testing for composeEmail()

Wrote a function to verify the construction of email messages in Go

* CW-44 Remove GmailAPI-related code

Removed commented out code since we are using the SMTP approach to send emails rather than using the full-fledged Gmail API

* Add a token to manipulate GitHub Secrets

* Revert "Add a token to manipulate GitHub Secrets"

This reverts commit 8c28fe69b88deaa2b74a0ed03ca8eaab84f5e9fc.

* Change Gmail API to MailJet API

We are using a third party mail provider to avoid having to store the attached email password in plaintext.

* Add Mailjet keys

These keys are attached to projects.website@csesoc.org.au

* Create global variables related to message bundles

Introduced the concept of bundling multiple enquiries into a single email to remain under MailJet's free tier limit of 200 emails/day.

* Create function that joins messages

joinMessages() returns a HTML string with all the messages concatenated

* Update mailing tests to match the new MailJet framework

Changed the tests to the new concatenated email body format.

* CW-58 Employ goroutines to run the timer concurrently

Goroutine is started on init and ticks every 15 minutes to send enquiry bundles

* CW-44 Fix bug to avoid sending empty bundles

* CW-44 Handle case where the Mailjet API can't be reached

If the API can't be reached rollover the pending enquiries to the next 15 minute cycle.

* CW-44 Fix reference error

* CW-44 Include safe stopping functionality to the echo server

The ability to stop gracefully lets us dispatch all pending enquiries before exiting the program.

* CW-44 Prepare branch to merge with dev

* CW-44 Merge feedback_test.go into mailing_test.go

* CW-44 Handle an echo server crash

By defering the dispatching of bundles I'm ensuring that they will always execute after main(), even if the echo server crashes.

* CW-44 Refactor shutdown procedure

Implemented a graceful shutdown as per Echo's documentation. Standardised the shutdown procedure to fuction with server errors and with the case of someone CTRL+C'ing the program (sending an interrupt signal), this new implementation works with both.
Used channels (https://gobyexample.com/channels) to communicate across goroutines.

* CW-44 Remove unnecessary goroutine creation

sendEnquiryBundle and sendFeedbackBundle can just run normally.

* CW-44 Change location of bundle size check

This was done to simplify the execution order by preventing the entering of send* functions if their corresponding bundles are empty.

* CW-44 Combine mailing timers into one goroutine with different intervals

This was done to simplify the program and reduce the number of concurrent goroutines.

* CW-44 Refactor mailing.go by merging two separate workflows

Created enums to define the 3 types of messages (info, sponsorship and feedback) and unify the way the server interfaces with mailing.go.
Merged the Handle* functions to take in both enquiries and feedbacks.
Refactored the send* functions into a single, reusable sendBundle() function.