CW-70 Update API documentation (#53)
* Create initial API framework

* CW-118 Refactored sponsors

* CW-118 Fix naming convention of API endpoints

- Added the API version to the endpoint prefix => /api/v1.
- Cleaned up server.go and fixed up comments
- Print progress of server startup

* CW-70 Add Swaggo modules and SwaggerUI endpoint

* CW-118 Fix trailing slash issue on URIs

The trailing slash on URIs was causing the tests to fail, as they were inconsistent throughout the backend server files.
e.g. /api/v1/faq != /api/v1/faq/

* CW-118 Break down API structure into groups

* restructured to APIClient and tested

* CW-70 Fix typo, missing slash

* CW-118 Unify sponsor endpoints

This new arrangement makes it intuitive to find things.

* CW-118 Remove 'middleware' function

Removed because it's not being used and it was implemented wrongly.

* CW-118 Fix naming inconsistencies

* CW-118 Fix URI path

* CW-118 Fix another typo

* Reverted local api url

* Fix Docker issue where the backend CompileDaemon fails

* CW-119 Add godoc for GetSponsor

Testing the doc generation for a single endpoint first.

* Remove Docker volumes from the backend

The server app was stored in a folder named '~' in the docker container. This change allows it to now be in the root directory.

* CW-125 Generate Swagger doc files dynamically in docker

* CW-125 Add doc files to the repo so CI works

The Dockerfile now removes the current docs directory and recreates when executed.

* CW-119 Add docstrings to Sponsor related API endpoints

* CW-119 Fix wrong failure message in documentation

* CW-119 Add docstrings to Mailing related API endpoints

To do this I had to split the HandleMessage() function into 3 separate handlers => one for each API endpoint.

* CW-70 Ensure every endpoint has intuitive error responses

Returning the error returned by the function that triggered the error isn't always useful.
In this commit I wrote descriptive error messages for all API endpoints.

* CW-119 Include responses and errors in documentation

Extended the Swagger documentation to include descriptive response and error messages.

* CW-119 Add docstrings to the FAQ API endpoint

* Prevent bundles from being dispatched while in development

* CW-119 Add Swagger integration for bearer tokens

Bearer tokens are now included in the header of certain sensitive API endpoints.

* CW-70 Add a comment on how to access SwaggerUI

* Fix Docker files to work with volumes

* CW-70 Refactor the handlers

Reverted back to the previous handling in one function while keeping separate API endpoints.

Co-authored-by: boboeee <jiechen4th@gmail.com>