Feat/about page (#63)
* CW-111 Add FAQs related to getting involved in subcommittees

I had to find a place to put information from the old CSESoc website related to subcommittees.

* CW-111 Add heading about Subcommittees

The Subcommittees heading includes a short paragraph on the role of subcommittees on the society and the short description of each society in the form of expansion panels.

* CW-111 Add Executive & Directors heading

The Executive & Directors are named in a table, structured similarly to the old website. This will change based on the wireframe for this page.

* CW-111 Remove unused data fields

* Remove trailing lines

* CW-111 Update information about the subcommittees

Using the Subcommittee Application Guide I added more information to the description of each subcommittee.

* CW-111 Add History heading

Used the content from the old website.

* CW-111 Add Workshop subcommittee to the list

* CW-111 Render Executives & Directors table dynamically from a JSON

This was done to reduce the number of lines of code and reduce the amount of 'dumb' HTML.

* CW-111 Simplify the rendering of the Executives & Directors table even more

* CW-111 Transcribbed all previous execs and directors in a JSON format

* CW-111 Display previous teams

I had to restructure the JSON to keep the order by using arrays instead of objects.

* CW-111 Update JWT's expiry time

JWTs expired yesterday. I added an extra 2 months to their expiry time.