Refactor fetch to be inside component.
Instead of having Home.vue pass the prop it is fetched directly in the
file. This is because it's hard to set default constant values fetched
from API because it is not fetched yet and doesn't re-render when it
does. Will ask what the team think or if there are other work arounds.