CW-154 Menu component (#69)
* CW-154 Add semitransparent overlay when clicking the menu toggle button

* CW-154 Embed overlay on the whole App

This makes the menu accessible from any page in the website.

* CW-154 Add links on the overlay

* CW-154 Connect links to existing pages

* CW-154 Auto-hide menu when a link is clicked

The status of the menu (shown or not) is controlled by the parent component (App.vue). So when a button is clicked we must $emit a signal to the parent to update the status of the menu to hidden.

* CW-154 Apply design elements relating to hovering on links

The menu links scale up x2 and become *bold* when hovered on, in their spots without interfering with other links.

* CW-154 Make menu elements reactive to the size change of other elements

* CW-154 Add automatic resizing based on window height

Menu links will resize automatically to ensure compatibility with any screen size.

* CW-154 Enable exiting the menu by clicking the background

* CW-154 Narrow the clicking area for the links

* CW-154 Make Menu component background fully black

* CW-154 Vanish Menu component when the CSESoc_logo is clicked

* CW-154 Convert `a` tags to `label` tags around `RouterLink` tags

To adhere to vue-router standards.

* CW-154 Disable scrolling when the Menu component is open

* CW-154 Change the order of the links

The new order is: About, Events, CSE Resources, Sponsors, Shop, Engage

* CW-154 Add link to Sponsors page

* CW-154 Fix bug that automatically scrolls to the top

This bug was introduced by disabling scrolling. It consisted of the page automatically resetting to the top when the scrolling was disabled/enabled.
Normally this would be worked around using JS and the window.scrollTo() but in Vue `window` is not in scope (outside mount and unmount).

* CW-154 Scroll to top for all route navigations

Otherwise the scroll position is saved across routes.